pyproject.toml
[project]
name = "naq"
version = "0.1.3"
description = "A simple and lightweight NATS queue worker"
readme = "README.md"
authors = [
    { name = "legout", email = "ligno.blades@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "nats-py>=2.6.0",
    "cloudpickle>=2.0.0",
    "typer>=0.9.0", # Keep typer[all] for CLI convenience
    "loguru>=0.7.0",
    "croniter>=1.4.0",
    "rich>=13.3.0", # For prettier CLI output (tables, etc.)
    "anyio>=4.9.0",
    "msgspec>=0.18.0",
]

[project.scripts]
naq = "naq.cli.main:app"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.5",
    "freezegun>=1.5.1",
]
dev = [
    "ipython>=9.0.2",
    "isort>=6.0.1",
    "nest-asyncio>=1.6.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.11.4",
    "six>=1.17.0",
]
dashboard = [
    "datastar-py>=0.6.3",
    "htmy>=0.7.3",
    "jinja2>=3.1.6",
    "sanic>=25.3.0",
    "sanic-ext>=24.12.0",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = "src"
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"  # Add this line

[dependency-groups]
dev = [
    "anyio>=4.9.0",
    "freezegun>=1.5.5",
    "ipython>=9.4.0",
    "nest-asyncio>=1.6.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "quarto>=0.1.0",
    "rq>=2.4.1",
]
test = [
    "freezegun>=1.5.4",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
]
